requirements.txt
django
djangorestframework
psycopg2-binary
redis
langgraph
fastapi
uvicorn
requests
streamlit